feat: getExamples() helper and general metadata file.
