+++ express-zod-api/src/date-in-schema.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { brandProperty } from "./brand";
+import { brandProperty } from "./metadata.ts";
 
 export const ezDateInBrand = Symbol("DateIn");
